refactor: extract NotificationSubscription for isolate-agnostic reuse
Previously the notification subscription lived inside
notification_provider.dart's _initializeAndListen and was gated on a
Riverpod Ref (so it could only run from the widget tree). This
prevents the foreground-task background isolate from reusing the same
subscription logic, which is what PR B needs.

Extract the subscription into lib/services/notification_subscription.dart
as a plain class. It takes dependencies via callbacks rather than Ref:

  - ActiveChatGetter: String? Function()
  - LocaleGetter: Locale Function()

The main-isolate provider wires these with ref.read(...). The upcoming
task handler will wire them with isolate-local defaults (no active
chat, locale from SharedPreferences).

Side effects of the extraction:
- formatNotification moves to the new file (still @visibleForTesting)
- handleNotificationUpdate becomes NotificationSubscription.handleUpdate
  (tests migrate accordingly)
- notification_provider.dart shrinks; its coverage:ignore block now
  wraps a smaller amount of genuinely-untestable Ref-wiring code

All ~40 existing tests for formatting + update handling migrate to
test/services/notification_subscription_test.dart. Coverage holds at
99%+. No behavioral change — this is a pure extraction.

Foundation for the next commit, which wires the new subscription into
the foreground-task handler for headless post-reboot delivery.
