+++ lib/providers/notification_provider.dart
@@ -1,10 +1,7 @@
-import 'dart:async';
 import 'dart:io' show Platform;
 
-import 'package:flutter/foundation.dart' show visibleForTesting;
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:logging/logging.dart';
-import 'package:whitenoise/l10n/generated/app_localizations.dart';
 import 'package:whitenoise/providers/active_chat_provider.dart';
 import 'package:whitenoise/providers/auth_provider.dart';
 import 'package:whitenoise/providers/foreground_service_provider.dart';
@@ -13,10 +10,7 @@ import 'package:whitenoise/routes.dart';
 import 'package:whitenoise/services/android_play_services_service.dart';
 import 'package:whitenoise/services/foreground_service.dart';
 import 'package:whitenoise/services/notification_service.dart';
-import 'package:whitenoise/services/user_service.dart';
-import 'package:whitenoise/src/rust/api/accounts.dart' as accounts_api;
-import 'package:whitenoise/src/rust/api/notifications.dart' as notifications_api;
-import 'package:whitenoise/utils/metadata.dart';
+import 'package:whitenoise/services/notification_subscription.dart';
 
 final _logger = Logger('NotificationProvider');
 
@@ -38,30 +32,35 @@ final notificationListenerProvider = Provider.autoDispose<void>((ref) {
   const androidPlayServicesService = AndroidPlayServicesService();
   final notificationService = ref.read(notificationServiceProvider);
   final foregroundService = ref.read(foregroundServiceProvider);
-  StreamSubscription<notifications_api.NotificationUpdate>? subscription;
+
+  final subscription = NotificationSubscription(
+    notificationService: notificationService,
+    getActiveChatId: () => ref.read(activeChatProvider),
+    getLocale: () => ref.read(localeProvider.notifier).resolveLocale(),
+  );
 
   ref.onDispose(() {
-    subscription?.cancel();
+    subscription.stop();
     foregroundService.stop();
     _logger.info('Notification listener disposed');
   });
 
-  _initializeAndListen(androidPlayServicesService, notificationService, foregroundService, ref, (
-    sub,
-  ) {
-    subscription = sub;
-  });
+  _startForegroundAndSubscribe(
+    androidPlayServicesService,
+    foregroundService,
+    subscription,
+    ref,
+  );
 });
 
-void _initializeAndListen(
-  AndroidPlayServicesService androidPlayServicesService,
-  NotificationService notificationService,
+Future<void> _startForegroundAndSubscribe(
+  AndroidPlayServicesService playServicesService,
   ForegroundService foregroundService,
+  NotificationSubscription subscription,
   Ref ref,
-  void Function(StreamSubscription<notifications_api.NotificationUpdate>) onSubscription,
 ) async {
   try {
-    final playServicesAvailability = await androidPlayServicesService.getAvailability();
+    final playServicesAvailability = await playServicesService.getAvailability();
     if (!ref.mounted) return;
     if (playServicesAvailability.isAvailable) {
       _logger.info(
@@ -75,11 +74,6 @@ void _initializeAndListen(
       );
     }
 
-    await notificationService.initialize();
-    if (!ref.mounted) return;
-    await notificationService.requestPermission();
-    if (!ref.mounted) return;
-
     await foregroundService.start();
     if (!ref.mounted) {
       await foregroundService.stop();
@@ -87,119 +81,18 @@ void _initializeAndListen(
     }
     await foregroundService.requestBatteryOptimizationExemption();
 
-    final stream = notifications_api.subscribeToNotifications();
-
-    final subscription = stream.listen(
-      (update) async {
-        try {
-          await handleNotificationUpdate(update, notificationService, ref);
-        } catch (error, stackTrace) {
-          _logger.severe('Error handling notification update', error, stackTrace);
-        }
-      },
-      onError: (error) {
-        _logger.severe('Notification stream error', error);
-      },
-      onDone: () {
-        _logger.info('Notification stream closed');
-      },
-    );
-
+    await subscription.start();
     if (!ref.mounted) {
-      await subscription.cancel();
+      await subscription.stop();
       await foregroundService.stop();
       return;
     }
-    onSubscription(subscription);
     _logger.info('Notification listener started');
   } catch (error, stackTrace) {
     _logger.severe('Failed to initialize notification listener', error, stackTrace);
   }
 }
-// coverage:ignore-end
 
-@visibleForTesting
-Future<void> handleNotificationUpdate(
-  notifications_api.NotificationUpdate update,
-  NotificationService notificationService,
-  Ref ref,
-) async {
-  final activeChat = ref.read(activeChatProvider);
-  if (activeChat == update.mlsGroupId) {
-    _logger.fine('Skipping notification for active chat ${update.mlsGroupId}');
-    return;
-  }
-
-  final locale = ref.read(localeProvider.notifier).resolveLocale();
-  final l10n = lookupAppLocalizations(locale);
-
-  final accounts = await accounts_api.getAccounts();
-  final String? receiverName = accounts.length > 1
-      ? (update.receiver.displayName ?? l10n.unknownUser)
-      : null;
-
-  final senderName = await _resolveSenderName(update.sender);
-
-  final (title, body, isInvite) = formatNotification(
-    update,
-    l10n,
-    receiverName: receiverName,
-    senderName: senderName,
-  );
-
-  await notificationService.show(
-    groupId: update.mlsGroupId,
-    title: title,
-    body: body,
-    receiverPubkey: update.receiver.pubkey,
-    isInvite: isInvite,
-  );
-}
-
-Future<String?> _resolveSenderName(notifications_api.NotificationUser sender) async {
-  if (sender.displayName != null) return sender.displayName;
-  try {
-    final metadata = await UserService(sender.pubkey).getInitialMetadata();
-    return presentName(metadata);
-  } catch (e) {
-    _logger.warning('Failed to fetch sender metadata', e);
-    return null;
-  }
-}
-
-@visibleForTesting
-(String title, String body, bool isInvite) formatNotification(
-  notifications_api.NotificationUpdate update,
-  AppLocalizations l10n, {
-  String? receiverName,
-  String? senderName,
-}) {
-  senderName ??= update.sender.displayName ?? l10n.unknownUser;
-
-  String applyReceiver(String title) {
-    if (receiverName == null) return title;
-    return '$title ($receiverName)';
-  }
-
-  switch (update.trigger) {
-    case notifications_api.NotificationTrigger.newMessage:
-      if (update.isDm) {
-        return (applyReceiver(senderName), update.content, false);
-      } else {
-        final groupName = update.groupName ?? l10n.unknownGroup;
-        return (applyReceiver(groupName), '$senderName: ${update.content}', false);
-      }
-    case notifications_api.NotificationTrigger.groupInvite:
-      if (update.isDm) {
-        return (applyReceiver(senderName), l10n.hasInvitedYouToSecureChat, true);
-      } else {
-        final groupName = update.groupName ?? l10n.unknownGroup;
-        return (applyReceiver(groupName), l10n.userInvitedYouToSecureChat(senderName), true);
-      }
-  }
-}
-
-// coverage:ignore-start
 Future<void> _onNotificationTap(
   Ref ref,
   String groupId,
